Next, rename the macros defined by the projects themselves.
Note that SODFG14A is currently *not* covered as expected. Also, getting
rid of the never-really-used GAMEVER_RESTORATION_WL6_APO11 is planned.

One trigger for the recent changes, is the fact that too long macros
(32 chars or more) may effectively be truncated by the compiler, and
this may even be considered valid behaviors by the C standard(s).

It's also a good chance to not just shorten macros, but also (usually)
shorten project file names by 1 char and use somewhat clearer names.