Finally, let's add Imagineer's Japanese localization (for DOS/V).
This is essentially the same as WL6GT14A.PRJ, although JAPAN is defined
instead of GOODTIMES. GAMEVER_WOLFREV could theoretically be the same
as well, but apparently the Japanese EXE was originally built
earlier, and the GOODTIMES-specific code changes were
probably not present in the codebase at all.

An appropriate SIGNON.OBJ file is also used, along with
a recreated GFXV_WJ6.H file.